requirements.txt
fastapi~=0.115.14
uvicorn
openai~=1.93.0
jinja2
weasyprint~=65.1
python-multipart
python-dotenv~=1.1.1
